requirements.txt
opencv-python-headless>=4.5.0        # OpenCV (обработка изображений, ORB, т.д.)
Pillow>=9.0.0                        # Для работы с изображениями в формате PIL (pHash)
numpy>=1.19.0                        # Массивы и операции с ними